+++ components/main/screens/Progress/components/calendar/Calendar.js
@@ -240,7 +240,7 @@ class MonthHolder extends React.Component {
         let month_timestamp_toString = new Date(this.props.data.year, this.props.data.month).getTime().toString()
         let month_chart_stats_map = Map(this.props.month_chart_stats)
 
-        return parseInt(month_chart_stats_map.getIn([month_timestamp_toString, "totalPoints"]))
+        return parseFloat(month_chart_stats_map.getIn([month_timestamp_toString, "totalPoints"]))
     }
 
     shouldComponentUpdate(nextProps, nextState) {
@@ -489,7 +489,7 @@ class WeekHolder extends React.Component {
     _calculateTotalPointsWeek = () => {
         let week_chart_stats_map = Map(this.props.week_chart_stats)
 
-        return parseInt(week_chart_stats_map.getIn([this.week_timestamp_toString, "totalPoints"]))
+        return parseFloat(week_chart_stats_map.getIn([this.week_timestamp_toString, "totalPoints"]))
     }
 
     render() {
@@ -583,7 +583,7 @@ class DayHolder extends React.Component {
     _calculateTotalPointsDay = () => {
         let day_chart_stats_map = Map(this.props.day_chart_stats)
 
-        return parseInt(day_chart_stats_map.getIn([this.day_timestamp_toString, "totalPoints"]))
+        return parseFloat(day_chart_stats_map.getIn([this.day_timestamp_toString, "totalPoints"]))
     }
 
     render() {
@@ -685,7 +685,7 @@ class UnchosenDayHolder extends React.Component {
     _calculateTotalPointsDay = () => {
         let day_chart_stats_map = Map(this.props.day_chart_stats)
 
-        return parseInt(day_chart_stats_map.getIn([this.day_timestamp_toString, "totalPoints"]))
+        return parseFloat(day_chart_stats_map.getIn([this.day_timestamp_toString, "totalPoints"]))
     }
 
     render() {
